+++ vmm/ci/restore-roundtrip.sh
@@ -40,7 +40,7 @@ finally:
 PY
 }
 
-RUST_LOG=info "$VMM" serve --socket "$SOCK" >"$LOG" 2>&1 &
+RUST_LOG=info "$VMM" serve --socket "$SOCK" --allow-unverified-restore >"$LOG" 2>&1 &
 SERVE_PID=$!
 sleep 1